feat: add amr-f approval mechanism
